WIP - Updating CLI to Click
