+++ app_template.yml
@@ -1,3 +1,3 @@
 name: ayna-app\s
 network: ayna-net0
-key: 000000000000000000000000\h\h\h\h\h\h\h\h
+key: \h\h\h\h\h\h\h\h
