+++ client/application/application_data.py
@@ -2,6 +2,7 @@ from dataclasses import dataclass
 from common.serializable import Serializable
 from typing import List
 from client.data_paths import base_dir, app_dir
+from common.file import file_helper
 
 
 @dataclass
@@ -25,3 +26,40 @@ class ApplicationData(Serializable):
         return f'Name: {self.name}\nKey: {self.key}\n' \
                f'Key Index: {self.key_index}\nNetwork: {self.network}\n' \
                f'Nodes: {self.nodes}'
+
+
+def app_name_list() -> list:
+    filenames = file_helper.list_files(base_dir + app_dir)
+    if not filenames:
+        return False
+
+    # remove file extension
+    filenames_fmt = []
+    for file in filenames:
+        filenames_fmt.append(file[:-4])
+
+    return filenames_fmt
+
+
+def app_key_list() -> list:
+    filenames = file_helper.list_files(base_dir + app_dir)
+    if not filenames:
+        return []
+
+    appkeys = []
+    for file in filenames:
+        app = ApplicationData.load(base_dir + app_dir + file)
+        appkeys.append(app.key)
+    return appkeys
+
+
+def app_key_index_list() -> list:
+    filenames = file_helper.list_files(base_dir + app_dir)
+    if not filenames:
+        return []
+
+    app_key_indexes = []
+    for file in filenames:
+        app = ApplicationData.load(base_dir + app_dir + file)
+        app_key_indexes.append(app.key_index)
+    return app_key_indexes
