+++ client/node/node_data.py
@@ -2,6 +2,7 @@ from dataclasses import dataclass
 from common.serializable import Serializable
 from typing import List
 from client.data_paths import base_dir, node_dir
+from common.file import file_helper
 
 
 @dataclass
@@ -28,3 +29,28 @@ class NodeData(Serializable):
                f'Device UUID: {self.device_uuid}\n' \
                f'Devkey: {self.devkey}\nNetwork: {self.network}\n' \
                f'Apps: {self.apps}'
+
+
+def node_name_list() -> list:
+    filenames = file_helper.list_files(base_dir + node_dir)
+    if not filenames:
+        return False
+
+    # remove file extension
+    filenames_fmt = []
+    for file in filenames:
+        filenames_fmt.append(file[:-4])
+
+    return filenames_fmt
+
+
+def node_addr_list() -> list:
+    filenames = file_helper.list_files(base_dir + node_dir)
+    if not filenames:
+        return []
+
+    node_addrs = []
+    for file in filenames:
+        node = NodeData.load(base_dir + node_dir + file)
+        node_addrs.append(node.addr)
+    return node_addrs
